PizzaOrderSystem
The project aims to have users design their own pizzas and add the user to the database after paying.
The Pizza Order system starts with the users choosing the pizza on the menu and the sauces they want.  They have to pay with their choice of sauce and pizza.  Users are required to pay by credit card.  Each pizza has a description and price.